make InputRefShifter a top level class
